+++ stylus-tools/precompiles/ArbOwnerPublic.sol
@@ -46,7 +46,7 @@ interface ArbOwnerPublic {
 
     /// @notice Get the max amount of stylus contract fragments that can be used to deploy a stylus contract
     /// @notice Available in ArbOS version 60 and above
-    function getMaxStylusContractFragments() external view returns (uint16);
+    function getMaxStylusContractFragments() external view returns (uint8);
 
     event ChainOwnerRectified(address rectifiedOwner);
 }
